+++ src/main/java/org/example/ssj3pj/controller/ContentsController.java
@@ -2,7 +2,7 @@ package org.example.ssj3pj.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.ssj3pj.dto.YoutubeContentDetailDto;
+import org.example.ssj3pj.dto.youtube.YoutubeContentDetailDto;
 import org.example.ssj3pj.services.ContentsService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
